Ebishrimp

+++ server/main.go
@@ -5,45 +5,90 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"os"
+
+	confparser "github.com/ebishrimp/config-file-parser-go"
 
 	_ "github.com/go-sql-driver/mysql"
 )
 
 var db *sql.DB
+var conf *confparser.Config
+var configFilePath string = "./fileserver.conf.template"
+
+// Basic configurations
+var ListenPort string
+var sqlUsername string
+var sqlPassword string
+
+// Request restrictions
+var allowUpload bool
+var allowDownload bool
+var allowOverwrite bool
+var allowDelete bool
+
+var whiteList bool
+
+// pseudo raid 0 settings
+var raid0 bool
+var raidpath string
 
 func main() {
-	http.HandleFunc("/test", testHandler)
+	configParse()
+	configLoad(conf)
+	fmt.Println("configs loaded successfully")
 
 	http.HandleFunc("/upload", uploadHandler)
 	http.HandleFunc("/download", downloadHandler)
 	http.HandleFunc("/overwrite", overWriteHandler)
 	http.HandleFunc("/delete", deleteHandler)
+	fmt.Println("Handlers registered successfully")
 
 	fmt.Println("Connecting to the database...")
 	dbConnect()
 	defer db.Close()
 
-	fmt.Println("Server is running on port 50080...")
-	http.ListenAndServe(":50080", nil)
-}
-
-func testHandler(w http.ResponseWriter, r *http.Request) {
-	if r.Method == "GET" || r.Method == "POST" {
-
-		name := r.URL.Query().Get("name")
-		hard := r.URL.Query().Get("hard")
-		app := r.URL.Query().Get("app")
-
-		if name == "" || hard == "" || app == "" {
-			w.Write([]byte("Missing parameters"))
-			return
+	fmt.Println("Server is running on port " + ListenPort + "...")
+	if err := http.ListenAndServe(":"+ListenPort, nil); err != nil {
+		log.Println("Error starting server on port " + ListenPort + ", starting on port 50080 instead")
+		if err := http.ListenAndServe(":50080", nil); err != nil {
+			log.Fatal(err)
 		}
 	}
+}
 
+func configParse() {
+	f, err := os.Open(configFilePath)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer f.Close()
+
+	configs, err := confparser.Parse(f)
+	if err != nil {
+		log.Fatal(err)
+	}
+	conf = configs
+}
+
+func configLoad(c *confparser.Config) {
+	ListenPort = c.GetValue("ListenPort")
+
+	sqlUsername = c.GetValue("mysqlUsername")
+	sqlPassword = c.GetValue("mysqlPassword")
+
+	allowUpload = c.GetValue("allowUpload") == "yes"
+	allowDownload = c.GetValue("allowDownload") == "yes"
+	allowOverwrite = c.GetValue("allowOverwrite") == "yes"
+	allowDelete = c.GetValue("allowDelete") == "yes"
+
+	raid0 = c.GetValue("raid0") == "yes"
+	raidpath = c.GetValue("raidpath")
+	whiteList = c.GetValue("whiteList") == "yes"
 }
 
 func dbConnect() {
-	data, err := sql.Open("mysql", "fileserver:fileserver@tcp(localhost:3306)/fileserver")
+	data, err := sql.Open("mysql", sqlUsername+":"+sqlPassword+"@tcp(localhost:3306)/fileserver")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -67,6 +112,11 @@ func uploadHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if !allowUpload {
+		http.Error(w, "Upload not allowed", http.StatusForbidden)
+		return
+	}
+
 	UploadOperation(db, name, hard, app, w)
 }
 
@@ -77,6 +127,11 @@ func downloadHandler(w http.ResponseWriter, r *http.Request) {
 
 	if name == "" || hard == "" || app == "" {
 		w.Write([]byte("Missing parameters"))
+		return
+	}
+
+	if !allowDownload {
+		http.Error(w, "Download not allowed", http.StatusForbidden)
 		return
 	}
 
@@ -112,6 +167,11 @@ func overWriteHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if !allowOverwrite {
+		http.Error(w, "Overwrite not allowed", http.StatusForbidden)
+		return
+	}
+
 	OverwriteOperation(db, name, hard, app)
 }
 
@@ -124,5 +184,11 @@ func deleteHandler(w http.ResponseWriter, r *http.Request) {
 		w.Write([]byte("Missing parameters"))
 		return
 	}
+
+	if !allowDelete {
+		http.Error(w, "Delete not allowed", http.StatusForbidden)
+		return
+	}
+
 	DeleteOperation(db, name, hard, app, w)
 }
